pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "demmaker"
version = "0.1.0"

[tool.poetry.dependencies]
pillow = "^8.3"
python = "^3.9"
typer = "^0.4"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
pylint = "^2.6.0"

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
